[Issue-1997] Split files
