+++ packages/extension-base/src/services/balance-service/helpers/substrate/index.ts
@@ -1,80 +1,25 @@
 // Copyright 2019-2022 @subwallet/extension-base
 // SPDX-License-Identifier: Apache-2.0
 
-import { SignedBalance } from '@equilab/api/genshiro/interfaces';
 import { _AssetType, _ChainAsset, _ChainInfo } from '@subwallet/chain-list/types';
 import { APIItemState, BalanceItem, TokenBalanceRaw } from '@subwallet/extension-base/background/KoniTypes';
-import { ASTAR_REFRESH_BALANCE_INTERVAL, SUB_TOKEN_REFRESH_BALANCE_INTERVAL } from '@subwallet/extension-base/constants';
+import { SUB_TOKEN_REFRESH_BALANCE_INTERVAL } from '@subwallet/extension-base/constants';
 import { PalletNominationPoolsPoolMember } from '@subwallet/extension-base/koni/api/staking/bonding/utils';
-import { getEVMBalance } from '@subwallet/extension-base/koni/api/tokens/evm/balance';
-import { getERC20Contract } from '@subwallet/extension-base/koni/api/tokens/evm/web3';
 import { getPSP22ContractPromise } from '@subwallet/extension-base/koni/api/tokens/wasm';
 import { getDefaultWeightV2 } from '@subwallet/extension-base/koni/api/tokens/wasm/utils';
 import { state } from '@subwallet/extension-base/koni/background/handlers';
-import { _BALANCE_CHAIN_GROUP, _MANTA_ZK_CHAIN_GROUP, _PURE_EVM_CHAINS, _ZK_ASSET_PREFIX } from '@subwallet/extension-base/services/chain-service/constants';
+import { subscribeERC20Interval } from '@subwallet/extension-base/services/balance-service/helpers/evm';
+import { _BALANCE_CHAIN_GROUP, _MANTA_ZK_CHAIN_GROUP, _ZK_ASSET_PREFIX } from '@subwallet/extension-base/services/chain-service/constants';
 import { _EvmApi, _SubstrateApi } from '@subwallet/extension-base/services/chain-service/types';
-import { _checkSmartContractSupportByChain, _getChainNativeTokenSlug, _getContractAddressOfToken, _getTokenOnChainAssetId, _getTokenOnChainInfo, _isChainEvmCompatible, _isPureEvmChain, _isSubstrateRelayChain } from '@subwallet/extension-base/services/chain-service/utils';
-import { categoryAddresses, sumBN } from '@subwallet/extension-base/utils';
-import BigN from 'bignumber.js';
-import { Contract } from 'web3-eth-contract';
+import { _checkSmartContractSupportByChain, _getChainNativeTokenSlug, _getContractAddressOfToken, _getTokenOnChainAssetId, _getTokenOnChainInfo, _isChainEvmCompatible, _isSubstrateRelayChain } from '@subwallet/extension-base/services/chain-service/utils';
+import { sumBN } from '@subwallet/extension-base/utils';
 
 import { ApiPromise } from '@polkadot/api';
 import { ContractPromise } from '@polkadot/api-contract';
 import { AccountInfo } from '@polkadot/types/interfaces';
 import { BN, BN_ZERO } from '@polkadot/util';
 
-type EqBalanceItem = [number, { positive: number }];
-type EqBalanceV0 = {
-  v0: {
-    lock: number,
-    balance: EqBalanceItem[]
-  }
-}
-
-// main subscription
-export function subscribeBalance (addresses: string[], chainInfoMap: Record<string, _ChainInfo>, substrateApiMap: Record<string, _SubstrateApi>, evmApiMap: Record<string, _EvmApi>, callback: (rs: BalanceItem) => void) {
-  const [substrateAddresses, evmAddresses] = categoryAddresses(addresses);
-
-  // Looping over each chain
-  const unsubList = Object.entries(chainInfoMap).map(async ([chainSlug, chainInfo]) => {
-    const useAddresses = _isChainEvmCompatible(chainInfo) ? evmAddresses : substrateAddresses;
-
-    if (_isPureEvmChain(chainInfo)) {
-      const nativeTokenInfo = state.getNativeTokenInfo(chainSlug);
-
-      return subscribeEVMBalance(chainSlug, useAddresses, evmApiMap, callback, nativeTokenInfo);
-    }
-
-    if (!useAddresses || useAddresses.length === 0 || _PURE_EVM_CHAINS.indexOf(chainSlug) > -1) {
-      const fungibleTokensByChain = state.chainService.getFungibleTokensByChain(chainSlug, true);
-      const now = new Date().getTime();
-
-      Object.values(fungibleTokensByChain).map((token) => {
-        return {
-          tokenSlug: token.slug,
-          free: '0',
-          locked: '0',
-          state: APIItemState.READY,
-          timestamp: now
-        } as BalanceItem;
-      }).forEach(callback);
-
-      return undefined;
-    }
-
-    const networkAPI = await substrateApiMap[chainSlug].isReady;
-
-    return subscribeSubstrateBalance(useAddresses, chainInfo, chainSlug, networkAPI, evmApiMap, callback);
-  });
-
-  return () => {
-    unsubList.forEach((subProm) => {
-      subProm.then((unsub) => {
-        unsub && unsub();
-      }).catch(console.error);
-    });
-  };
-}
+import { subscribeEqBalanceAccountPallet, subscribeEquilibriumTokenBalance } from './equilibrium';
 
 export async function subscribeSubstrateBalance (addresses: string[], chainInfo: _ChainInfo, chain: string, networkAPI: _SubstrateApi, evmApiMap: Record<string, _EvmApi>, callBack: (rs: BalanceItem) => void) {
   let unsubNativeToken: () => void;
@@ -186,50 +131,6 @@ async function subscribeWithSystemAccountPallet (addresses: string[], chainInfo:
   };
 }
 
-function subscribeERC20Interval (addresses: string[], chain: string, evmApiMap: Record<string, _EvmApi>, callBack: (result: BalanceItem) => void): () => void {
-  let tokenList = {} as Record<string, _ChainAsset>;
-  const erc20ContractMap = {} as Record<string, Contract>;
-
-  const getTokenBalances = () => {
-    Object.values(tokenList).map(async (tokenInfo) => {
-      let free = new BN(0);
-
-      try {
-        const contract = erc20ContractMap[tokenInfo.slug];
-        const balanceList = await Promise.all(addresses.map((address): Promise<string> => {
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-return,@typescript-eslint/no-unsafe-call,@typescript-eslint/no-unsafe-member-access
-          return contract.methods.balanceOf(address).call();
-        }));
-
-        free = sumBN(balanceList.map((balance) => new BN(balance || 0)));
-
-        callBack({
-          tokenSlug: tokenInfo.slug,
-          free: free.toString(),
-          locked: '0',
-          state: APIItemState.READY
-        } as BalanceItem);
-      } catch (err) {
-        console.log(tokenInfo.slug, err);
-      }
-    });
-  };
-
-  tokenList = state.getAssetByChainAndAsset(chain, [_AssetType.ERC20]);
-
-  Object.entries(tokenList).forEach(([slug, tokenInfo]) => {
-    erc20ContractMap[slug] = getERC20Contract(chain, _getContractAddressOfToken(tokenInfo), evmApiMap);
-  });
-
-  getTokenBalances();
-
-  const interval = setInterval(getTokenBalances, SUB_TOKEN_REFRESH_BALANCE_INTERVAL);
-
-  return () => {
-    clearInterval(interval);
-  };
-}
-
 function subscribePSP22Balance (addresses: string[], chain: string, api: ApiPromise, callBack: (result: BalanceItem) => void) {
   let tokenList = {} as Record<string, _ChainAsset>;
   const psp22ContractMap = {} as Record<string, ContractPromise>;
@@ -272,79 +173,6 @@ function subscribePSP22Balance (addresses: string[], chain: string, api: ApiProm
 
   return () => {
     clearInterval(interval);
-  };
-}
-
-async function subscribeEquilibriumTokenBalance (addresses: string[], chain: string, api: ApiPromise, callBack: (rs: BalanceItem) => void, includeNativeToken?: boolean): Promise<() => void> {
-  const tokenTypes = includeNativeToken ? [_AssetType.NATIVE, _AssetType.LOCAL] : [_AssetType.LOCAL];
-  const tokenMap = state.getAssetByChainAndAsset(chain, tokenTypes);
-
-  const unsub = await api.query.system.account.multi(addresses, (balances: Record<string, any>[]) => { // Equilibrium customizes the SystemAccount pallet
-    Object.values(tokenMap).forEach((tokenInfo) => {
-      const assetId = _getTokenOnChainAssetId(tokenInfo);
-      let tokenFreeBalance = BN_ZERO;
-
-      for (const balance of balances) {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-argument,@typescript-eslint/no-unsafe-call,@typescript-eslint/no-unsafe-member-access
-        const balancesData = JSON.parse(balance.data.toString()) as EqBalanceV0;
-        const balanceList = balancesData.v0.balance;
-
-        // @ts-ignore
-        const freeTokenBalance = balanceList.find((data: EqBalanceItem) => data[0] === parseInt(assetId));
-        const bnFreeTokenBalance = freeTokenBalance ? new BN(new BigN(freeTokenBalance[1].positive).toString()) : BN_ZERO;
-
-        tokenFreeBalance = tokenFreeBalance.add(bnFreeTokenBalance);
-      }
-
-      const tokenBalance: BalanceItem = {
-        free: tokenFreeBalance.toString(),
-        locked: '0', // Equilibrium doesn't show locked balance
-        state: APIItemState.READY,
-        tokenSlug: tokenInfo.slug
-      };
-
-      callBack(tokenBalance);
-    });
-  });
-
-  return () => {
-    unsub();
-  };
-}
-
-// eslint-disable-next-line @typescript-eslint/require-await
-async function subscribeEqBalanceAccountPallet (addresses: string[], chain: string, api: ApiPromise, callBack: (rs: BalanceItem) => void, includeNativeToken?: boolean): Promise<() => void> {
-  const tokenTypes = includeNativeToken ? [_AssetType.NATIVE, _AssetType.LOCAL] : [_AssetType.LOCAL];
-  const tokenMap = state.getAssetByChainAndAsset(chain, tokenTypes);
-
-  const unsubList = Object.values(tokenMap).map(async (tokenInfo) => {
-    try {
-      const assetId = _getTokenOnChainAssetId(tokenInfo);
-      const unsub = await api.query.eqBalances.account.multi(addresses.map((address) => [address, [assetId]]), (balances: SignedBalance[]) => {
-        const tokenBalance = {
-          free: sumBN(balances.map((b) => (b.asPositive))).toString(),
-          locked: '0', // Equilibrium doesn't show locked balance
-          state: APIItemState.READY,
-          tokenSlug: tokenInfo.slug
-        };
-
-        callBack(tokenBalance);
-      });
-
-      return unsub;
-    } catch (err) {
-      console.warn(err);
-
-      return undefined;
-    }
-  });
-
-  return () => {
-    unsubList.forEach((subProm) => {
-      subProm.then((unsub) => {
-        unsub && unsub();
-      }).catch(console.error);
-    });
   };
 }
 
@@ -457,33 +285,5 @@ async function subscribeAssetsAccountPallet (addresses: string[], chain: string,
     unsubList.forEach((unsub) => {
       unsub && unsub();
     });
-  };
-}
-
-export function subscribeEVMBalance (chain: string, addresses: string[], evmApiMap: Record<string, _EvmApi>, callback: (rs: BalanceItem) => void, tokenInfo: _ChainAsset) {
-  const balanceItem = {
-    tokenSlug: tokenInfo.slug,
-    state: APIItemState.PENDING,
-    free: '0',
-    locked: '0'
-  } as BalanceItem;
-
-  function getBalance () {
-    getEVMBalance(chain, addresses, evmApiMap)
-      .then((balances) => {
-        balanceItem.free = sumBN(balances.map((b) => (new BN(b || '0')))).toString();
-        balanceItem.state = APIItemState.READY;
-        callback(balanceItem);
-      })
-      .catch(console.warn);
-  }
-
-  getBalance();
-  const interval = setInterval(getBalance, ASTAR_REFRESH_BALANCE_INTERVAL);
-  const unsub2 = subscribeERC20Interval(addresses, chain, evmApiMap, callback);
-
-  return () => {
-    clearInterval(interval);
-    unsub2 && unsub2();
   };
 }
